move environment variables related utilities for package config
